Bump to go1.16 (#214)
